+++ src/Maui/Features/Creatures/Import/Create.razor.cs
@@ -1,7 +1,6 @@
-﻿using TabletopPowerTools.Core.Features.Creatures.Import;
-using TabletopPowerTools.Core.Models;
+﻿using TabletopPowerTools.Core.Models;
 using TabletopPowerTools.Core.Models.ViewModels;
-
+using TabletopPowerTools.Core.Features.Creatures;
 
 namespace TabletopPowerTools.Maui.Features.Creatures.Import;
 public partial class Create
